Put all the data in a separated file
